// backend/src/types/ClanMembers.ts
export interface Member {
    id: number;
    clan_id: number;
    nickname: string;
    name: string;
    pubg_id: string;
    age: number;
    telegram_tag: string;
    city: string | null;
    clan: number;
    actor_id: number | null;
    kd: string | null;
    created_at: string
   // created_at: Date;
}